separando metodos para a classe post
