+++ modulo5/Projeto-labook/src/business/PostBusiness.ts
@@ -4,6 +4,7 @@ import { Post } from "../model/Post";
 import { Authenticator } from "../services/Authenticator";
 import { IdGeneration } from "../services/IdGeneration";
 import { postDTO } from "../types/DTO";
+import { likeInfo } from "../types/like";
 
 export class PostBusiness {
   private idGeneration: IdGeneration;
@@ -59,5 +60,44 @@ export class PostBusiness {
     }
     const post = await this.postData.findPostById(id);
     return post;
+  };
+  listOfPosts = async (token: string): Promise<Post[]> => {
+    if (!token) {
+      throw new Error("É necessário passar o token de acesso");
+    }
+    const tokenData = this.authenticator.getTokenData(token);
+    if (!tokenData) {
+      throw new Error("Usuário deslogado");
+    }
+
+    const getFriends = await this.postData.friends(tokenData.id);
+
+    const getAllPosts = await this.postData.returnPosts(getFriends.friend_id);
+
+    return getAllPosts;
+  };
+
+  listOfPostsByType = async (token: string): Promise<Post[]> => {
+    if (!token) throw new Error("É preciso passar um token de acesso");
+    const tokenData = this.authenticator.getTokenData(token);
+    if (!tokenData) throw new Error("Usuário deslogado");
+
+    const getFriends = await this.postData.friends(tokenData.id);
+    const allPostsByType = await this.postData.returnPostsByType(
+      getFriends.friend_id
+    );
+    return allPostsByType;
+  };
+
+  likingPost = async (info: likeInfo): Promise<string> => {
+    const { id, token } = info;
+    if (!id) throw new Error("É necessário passar um token para curtir o post");
+    if (!token) throw new Error("É preciso passar um token de acesso");
+
+    const tokenData = this.authenticator.getTokenData(token);
+    if (!tokenData) throw new Error("Usário Deslogado");
+
+    // const getPost = this.us
+    return "ellenzinha do meu coração, te amo mil e milhão";
   };
 }
