+++ modulo5/Projeto-labook/src/controller/UserController.ts
@@ -2,6 +2,7 @@ import { Request, Response } from "express";
 import { UserBusiness } from "../business/UserBusiness";
 import { Post } from "../model/Post";
 import { followUserDTO, loginDTO, signupDto } from "../types/DTO";
+import { likeInfo } from "../types/like";
 
 export class UserController {
   constructor(private userBusiness: UserBusiness) {}
@@ -83,37 +84,5 @@ export class UserController {
         res.status(500).send({ message: "Erro ao se conectar com o servidor" });
       }
     }
-  };
-
-  feed = async (req: Request, res: Response): Promise<Post[]> => {
-    try {
-      const token = req.headers.authorization;
-      const posts: Post[] = await this.userBusiness.listOfPosts(token);
-      res.status(200).send({ posts });
-      return posts;
-    } catch (error) {
-      if (error instanceof Error) {
-        res.status(400).send({ message: error.message });
-      } else {
-        res.status(500).send({ message: "Erro ao se conectar com o servidor" });
-      }
-    }
-  };
-
-  feedByType = async (req: Request, res: Response): Promise<Post[]> => {
-    try {
-      const token = req.headers.authorization;
-      const newFeed: Post[] = await this.userBusiness.listOfPostsByType(token);
-      res.status(200).send({ newFeed });
-      return newFeed;
-    } catch (error) {
-      if (error instanceof Error) {
-        res.status(400).send({ message: error.message });
-      } else {
-        res.status(500).send({ message: "Erro ao se conectar com o servidor" });
-      }
-    }
-
-    return;
   };
 }
