+++ modulo5/Projeto-labook/src/index.ts
@@ -19,6 +19,6 @@ app.post("/users/login", userController.login);
 app.post("/unfollow", userController.unfollowUser);
 app.post("/users/follow", userController.followUser);
 app.post("/post", postController.createPost);
-app.get("/feed/types", userController.feedByType);
-app.get("/feed", userController.feed);
+app.get("/feed/types", postController.feedByType);
+app.get("/feed", postController.feed);
 app.get("/post/:id", postController.getPostById);
