+++ modulo5/Projeto-labook/src/model/IUserData.ts
@@ -8,7 +8,7 @@ export interface IUserData {
   insertNewFriend(follow: followUserDTO): void;
   findFollowId(friend_id: string): Promise<boolean>;
   removeFollow(friend_id: string): Promise<void>;
-  friends(user_id: string): Promise<any>;
-  returnPosts(friend_id: string): Promise<Post[]>;
-  returnPostsByType(user_id: string): Promise<Post[]>;
+
+  // returnPosts(friend_id: string): Promise<Post[]>;
+  // returnPostsByType(user_id: string): Promise<Post[]>;
 }
