+++ modulo5/Projeto-labook/src/types/friend.ts
@@ -1,0 +1,4 @@
+export type friend = {
+  user_id: string;
+  friend_id: string;
+};
