+++ modulo5/Projeto-labook/src/types/like.ts
@@ -1,0 +1,4 @@
+export type likeInfo = {
+  id: string;
+  token: string;
+};
